service.is_running
service.is_running doesn't play well with upstart services as the
`service` command returns successfully no mater if the service is
running or not. This updates service.is_running to determine whether the
service is an upstart service or not by testing for the presence of an
upstart config at /etc/init/<service>.conf. If one exists it checks the
output of the status command for the word 'running', otherwise, it
returns whether the call to status succeeded or not.