Make /tmp the default temp dir for require.file() [#218]
When use_sudo is enabled in Fabric's put() function, it will first
upload the file to a temporary directory, before moving it as root.
The default is to use the user's home directory. However when it is
mounted over NFS with the root_squash option (the default), then root
will not be able to read the file, causing the move to fail.

We now default to /tmp to avoid this problem.