Tests: patch Fabric so that it accesses the real stdin
The py.test capture plugin replaces sys.stdin with a fake one.
This is usually good, as well-behaved tests should be unattended
and should not try to read from standard input.

Unfortunately, this causes Fabric to fail when it tries to setup
its input loop, so we monkey patch the real stdin back in.