Fix tomcat module on Debian 8 (jessie) [#293]
When installing from source, the code installs a custom init script
in /etc/init.d, but starts and stops the service using a higher-level
service management API, which fails on the systemd-based Debian 8.
So now we just call our init script directly.

Also, we now need to include LSB headers in the init script.

(Note that Tomcat 7 is available in Debian 8 so you'd probably better
use this package anyway.)