Tests: use generative env specs introduced in tox 1.8
This allows to run tests on a combination of Python versions
and target operating systems.

For a start, the following target operating systems are included
in the tox config:
- Debian 8 : official Debian box (debian/jessie64)
- Ubuntu 12.04 LTS : standard Vagrant box (hashicorp/precise64)
- Ubuntu 14.04 LTS : official Ubuntu box (ubuntu/trusty64)

Others will be added later, so feel free to suggest additions.

Also, starting with tox 2.0 we need to explicitly list environment
variables that need to be passed to the test commands.