Use curl instead of wget to download Oracle JDK [#114, #222]
Version 1.12 of `wget` (included in Debian 6) has a bug in the way
it handles SSL certificates, which causes verification to fail.

See: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=409938

This explains why some people needed to add the insecure
`--no-check-certificate` option to make it work.

Switching to curl in order to avoid this problem.